restructure some stuff
